av1: CB2 fix hdr10p stream without hdr10p infos. [1/1]
PD#SWPL-86046

Problem:
The decoding driver does not output the hdr10p flag to the
backend, resulting in the backend not getting the hdr10p data.

Solution:
When the relevant information of hdr10p is detected in the
decoding driver, the corresponding flag is passed to the backend.

Verify:
ohm

Change-Id: I06b275617eb9195204e667b9389e0be9f50e30e9
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>