h264: CB2 fix crash problem caused by abnormal afd data address. [1/1]
PD#SWPL-96491

Problem:
Because of the weak signal, some frames in the ucode parse out dirty sei
information. When a decoding exception reset occurs, the address where
the sei data is stored will be free, which will cause the backend to get
the abnormal address and cause a crash.

Solution:
The decoded afd information is updated to the backend every time it is decoded.

Verify:
T5W

Change-Id: Ibb5f19778964ccaffee945661677aeb02254005b
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>